+++ index.ts
@@ -1,12 +1,1 @@
-/**
- * Angular library starter.
- * Build a library compatible with Angular, AoT compilation & Tree shaking.
- * Written by Roberto Simonetti.
- * MIT license.
- * https://github.com/robisim74/angular-library-starter
- */
-
-/**
- * Entry point for all public APIs of the package.
- */
-export * from './src/angular-library-starter';
+export * from './public_api';