+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Angular library starter Changelog
 
+<a name="Mar 25, 2017"></a>
+### Mar 25, 2017
+* Upgrade to Angular 4 configuration
+
 <a name="Mar 6, 2017"></a>
 ### Mar 6, 2017
 * Add _compodoc_ for generating documentation